Ta i bruk useEøs på riktig måte og fjern Provider + rename context og flytt inn i eøs-mappe
