+++ src/frontend/context/UtenlandskPeriodeBeløp/UtenlandskPeriodeBeløpSkjemaContext.tsx
@@ -6,6 +6,10 @@ import { feil, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
 import type { Ressurs } from '@navikt/familie-typer';
 import { byggTomRessurs, RessursStatus } from '@navikt/familie-typer';
 
+import {
+    konverterDesimalverdiTilSkjemaVisning,
+    konverterSkjemaverdiTilDesimal,
+} from '../../komponenter/Fagsak/Behandlingsresultat/Eøs/useEøs';
 import type { IBehandling } from '../../typer/behandling';
 import type { OptionType } from '../../typer/common';
 import type {
@@ -18,10 +22,6 @@ import type { IIsoMånedPeriode } from '../../utils/dato';
 import { nyIsoMånedPeriode } from '../../utils/dato';
 import { erBarnGyldig, erEøsPeriodeGyldig, isEmpty, isNumeric } from '../../utils/eøsValidators';
 import { useBehandling } from '../behandlingContext/BehandlingContext';
-import {
-    konverterDesimalverdiTilSkjemaVisning,
-    konverterSkjemaverdiTilDesimal,
-} from '../Eøs/EøsContext';
 
 const erBeløpGyldig = (felt: FeltState<string | undefined>): FeltState<string | undefined> => {
     if (!felt.verdi || isEmpty(felt.verdi) || typeof felt.verdi != 'string') {
