+++ src/frontend/context/Valutakurs/ValutakursSkjemaContext.tsx
@@ -8,6 +8,10 @@ import { feil, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
 import type { Ressurs } from '@navikt/familie-typer';
 import { byggTomRessurs, RessursStatus } from '@navikt/familie-typer';
 
+import {
+    konverterDesimalverdiTilSkjemaVisning,
+    konverterSkjemaverdiTilDesimal,
+} from '../../komponenter/Fagsak/Behandlingsresultat/Eøs/useEøs';
 import type { IBehandling } from '../../typer/behandling';
 import type { OptionType } from '../../typer/common';
 import type { EøsPeriodeStatus, IRestValutakurs, IValutakurs } from '../../typer/eøsPerioder';
@@ -27,10 +31,6 @@ import {
     tellAntallDesimaler,
 } from '../../utils/eøsValidators';
 import { useBehandling } from '../behandlingContext/BehandlingContext';
-import {
-    konverterDesimalverdiTilSkjemaVisning,
-    konverterSkjemaverdiTilDesimal,
-} from '../Eøs/EøsContext';
 
 const erValutakursGyldig = (
     felt: FeltState<string | undefined>,
