+++ src/frontend/komponenter/Fagsak/BehandlingRouter.tsx
@@ -12,7 +12,6 @@ import { SammensattKontrollsakProvider } from './Vedtak/SammensattKontrollsak/us
 import Vilkårsvurdering from './Vilkårsvurdering/Vilkårsvurdering';
 import { useBehandling } from '../../context/behandlingContext/BehandlingContext';
 import { VedtakStegProvider } from '../../context/behandlingContext/useVedtakSteg';
-import { EøsProvider } from '../../context/Eøs/EøsContext';
 import { InstitusjonProvider } from '../../context/InstitusjonContext';
 import { SimuleringProvider } from '../../context/SimuleringContext';
 import { SøknadProvider } from '../../context/SøknadContext';
@@ -74,9 +73,7 @@ const BehandlingRouter: React.FC<Props> = ({ bruker }) => {
                 path="/tilkjent-ytelse"
                 element={
                     <TidslinjeProvider>
-                        <EøsProvider åpenBehandling={behandling}>
-                            <Behandlingsresultat åpenBehandling={behandling} />
-                        </EøsProvider>
+                        <Behandlingsresultat åpenBehandling={behandling} />
                     </TidslinjeProvider>
                 }
             />
