+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Eøs/EøsKomponenter/EøsSkjemaKomponenter.tsx
@@ -11,13 +11,13 @@ import {
     ASurfaceAction,
 } from '@navikt/ds-tokens/dist/tokens';
 
-import { mapEøsPeriodeStatusTilStatus } from '../../../../../context/Eøs/EøsContext';
 import StatusIkon from '../../../../../ikoner/StatusIkon';
 import { EøsPeriodeStatus, Vurderingsform } from '../../../../../typer/eøsPerioder';
 import type { IGrunnlagPerson } from '../../../../../typer/person';
 import { Datoformat, isoMånedPeriodeTilFormatertString } from '../../../../../utils/dato';
 import type { IIsoMånedPeriode } from '../../../../../utils/dato';
 import { lagPersonLabel } from '../../../../../utils/formatter';
+import { mapEøsPeriodeStatusTilStatus } from '../useEøs';
 
 interface IEøsPeriodeSkjemaContainerProps {
     $lesevisning: boolean;
